+++ spec/truffle/posix_spec.rb
@@ -8,11 +8,11 @@
 
 require_relative '../ruby/spec_helper'
 
-describe "Trffle::POSIX returns the correct value for an identity function returning" do
+describe "Truffle::POSIX returns the correct value for an identity function returning" do
   before :all do
     src = fixture __FILE__, "libtestnfi.c"
-    lib = src[0...-1] + RbConfig::CONFIG['NATIVE_DLEXT']
-    unless system "clang", "-shared", "-o", lib, src
+    lib = src[0...-1] + RbConfig::CONFIG['SOEXT']
+    unless system RbConfig::CONFIG['CC'], "-shared", "-o", lib, src
       abort "Could not compile libtestnfi"
     end