+++ test/truffle/cexts/xopenssl/ext/xopenssl/extconf.rb
@@ -4,9 +4,9 @@ require 'truffle/openssl-prefix'
 
 if ENV['OPENSSL_PREFIX']
   $CFLAGS += " -I #{ENV['OPENSSL_PREFIX']}/include"
-  $LIBS += " -l #{ENV['OPENSSL_PREFIX']}/lib/libssl.#{RbConfig::CONFIG['NATIVE_DLEXT']}"
+  $LIBS += " #{ENV['OPENSSL_PREFIX']}/lib/libssl.#{RbConfig::CONFIG['SOEXT']} #{ENV['OPENSSL_PREFIX']}/lib/libcrypto.#{RbConfig::CONFIG['SOEXT']}"
 else
-  $LIBS += " -l libssl.#{RbConfig::CONFIG['NATIVE_DLEXT']}"
+  $LIBS += " -lssl -lcrypto"
 end
 
 create_makefile('xopenssl')